Module Protocol.Seed_storage

type Tezos_protocol_environment_005_PsBABY5H__Environment.Error_monad.error +=

| Unknown of {

oldest : Tezos_raw_protocol_005_PsBABY5H.Cycle_repr.t;

cycle : Tezos_raw_protocol_005_PsBABY5H.Cycle_repr.t;

latest : Tezos_raw_protocol_005_PsBABY5H.Cycle_repr.t;

}

val init : Tezos_raw_protocol_005_PsBABY5H.Raw_context.t -> (Tezos_raw_protocol_005_PsBABY5H.Raw_context.t, Tezos_protocol_environment_005_PsBABY5H__Environment.Error_monad.error list) Stdlib.result Lwt.t

Generates the first preserved_cycles+2 seeds for which there are no nonces.

val for_cycle : Tezos_raw_protocol_005_PsBABY5H.Raw_context.t -> Tezos_raw_protocol_005_PsBABY5H.Cycle_repr.t -> (Tezos_raw_protocol_005_PsBABY5H.Seed_repr.seed, Tezos_protocol_environment_005_PsBABY5H__Environment.Error_monad.error list) Stdlib.result Lwt.t

val cycle_end : Tezos_raw_protocol_005_PsBABY5H.Raw_context.t -> Tezos_raw_protocol_005_PsBABY5H.Cycle_repr.t -> (Tezos_raw_protocol_005_PsBABY5H.Raw_context.t * Tezos_raw_protocol_005_PsBABY5H.Nonce_storage.unrevealed list, Tezos_protocol_environment_005_PsBABY5H__Environment.Error_monad.error list) Stdlib.result Lwt.t

If it is the end of the cycle, computes and stores the seed of cycle at distance preserved_cycle+2 in the future using the seed of the previous cycle and the revelations of the current one.