Module Storage.Active_delegates_with_rolls

Set of all active delegates with rolls.

type t

type context = t

type elt = Tezos_protocol_environment_005_PsBABY5H__Environment.Signature.Public_key_hash.t

The type of elements.

val mem : context -> elt -> bool Tezos_protocol_environment_005_PsBABY5H__Environment.Lwt.t

Tells if a elt is a member of the set

val add : context -> elt -> Raw_context.t Tezos_protocol_environment_005_PsBABY5H__Environment.Lwt.t

Adds a elt is a member of the set

val del : context -> elt -> Raw_context.t Tezos_protocol_environment_005_PsBABY5H__Environment.Lwt.t

Removes a elt of the set ; does nothing if not a member

val set : context -> elt -> bool -> Raw_context.t Tezos_protocol_environment_005_PsBABY5H__Environment.Lwt.t

Adds/Removes a elt of the set

val elements : context -> elt list Tezos_protocol_environment_005_PsBABY5H__Environment.Lwt.t

Returns the elements of the set, deserialized in a list in no particular order.

val fold : context -> init:'a -> f:(elt -> 'a -> 'a Tezos_protocol_environment_005_PsBABY5H__Environment.Lwt.t) -> 'a Tezos_protocol_environment_005_PsBABY5H__Environment.Lwt.t

Iterates over the elements of the set.

val clear : context -> Raw_context.t Tezos_protocol_environment_005_PsBABY5H__Environment.Lwt.t

Removes all elements in the set