Module Storage.Vote

module Current_period_kind : Storage_sigs.Single_data_storage with type value = Voting_period_repr.kind and type t := Raw_context.t

module Current_quorum_004 : Storage_sigs.Single_data_storage with type value = int32 and type t := Raw_context.t

Only for migration from 004. Expected quorum, in centile of percentage

module Participation_ema : Storage_sigs.Single_data_storage with type value = int32 and type t := Raw_context.t

Participation exponential moving average, in centile of percentage

module Current_proposal : Storage_sigs.Single_data_storage with type value = Tezos_protocol_environment_005_PsBABY5H__Environment.Protocol_hash.t and type t := Raw_context.t

module Listings_size : Storage_sigs.Single_data_storage with type value = int32 and type t := Raw_context.t

Sum of all rolls of all delegates.

module Listings : Storage_sigs.Indexed_data_storage with type key = Tezos_protocol_environment_005_PsBABY5H__Environment.Signature.Public_key_hash.t and type value = int32 and type t := Raw_context.t

Contains all delegates with their assigned number of rolls.

module Proposals : Storage_sigs.Data_set_storage with type elt = Tezos_protocol_environment_005_PsBABY5H__Environment.Protocol_hash.t * Tezos_protocol_environment_005_PsBABY5H__Environment.Signature.Public_key_hash.t and type t := Raw_context.t

Set of protocol proposal with corresponding proposer delegate

module Proposals_count : Storage_sigs.Indexed_data_storage with type key = Tezos_protocol_environment_005_PsBABY5H__Environment.Signature.Public_key_hash.t and type value = int and type t := Raw_context.t

Keeps for each delegate the number of proposed protocols

module Ballots : Storage_sigs.Indexed_data_storage with type key = Tezos_protocol_environment_005_PsBABY5H__Environment.Signature.Public_key_hash.t and type value = Vote_repr.ballot and type t := Raw_context.t

Contains for each delegate its ballot