Module type Storage_sigs.Non_iterable_indexed_data_storage

Restricted version of Indexed_data_storage w/o iterators.

type t

type context = t

type key

An abstract type for keys

type value

The type of values

val mem : context -> key -> bool Lwt.t

Tells if a given key is already bound to a storage bucket

val get : context -> key -> (value, Tezos_protocol_environment_005_PsBABY5H__Environment.Error_monad.error list) Stdlib.result Lwt.t

Retrieve a value from the storage bucket at a given key ; returns Storage_error Missing_key if the key is not set ; returns Storage_error Corrupted_data if the deserialisation fails.

val get_option : context -> key -> (value option, Tezos_protocol_environment_005_PsBABY5H__Environment.Error_monad.error list) Stdlib.result Lwt.t

Retrieve a value from the storage bucket at a given key ; returns None if the value is not set ; returns Storage_error Corrupted_data if the deserialisation fails.

val set : context -> key -> value -> (Raw_context.t, Tezos_protocol_environment_005_PsBABY5H__Environment.Error_monad.error list) Stdlib.result Lwt.t

Updates the content of a bucket ; returns A Storage_Error Missing_key if the value does not exists.

val init : context -> key -> value -> (Raw_context.t, Tezos_protocol_environment_005_PsBABY5H__Environment.Error_monad.error list) Stdlib.result Lwt.t

Allocates a storage bucket at the given key and initializes it ; returns a Storage_error Existing_key if the bucket exists.

val init_set : context -> key -> value -> Raw_context.t Lwt.t

Allocates a storage bucket at the given key and initializes it with a value ; just updates it if the bucket exists.

val set_option : context -> key -> value option -> Raw_context.t Lwt.t

When the value is Some v, allocates the data and initializes it with v ; just updates it if the bucket exists. When the valus is None, delete the storage bucket when the value ; does nothing if the bucket does not exists.

val delete : context -> key -> (Raw_context.t, Tezos_protocol_environment_005_PsBABY5H__Environment.Error_monad.error list) Stdlib.result Lwt.t

Delete a storage bucket and its contents ; returns a Storage_error Missing_key if the bucket does not exists.

val remove : context -> key -> Raw_context.t Lwt.t

Removes a storage bucket and its contents ; does nothing if the bucket does not exists.